navigator as component
