+++ README.md
@@ -1,4 +1,8 @@
 Yun's Portfolio
 
 - contact page photo
+
 - mobile version mq
+- navigator
+- overall padding
+- projectdetail
